Extract `Guide.Instruction` to the outer scope
