+++ DevToolkt/core/src/commonMain/kotlin/dev/toolkt/core/data_structures/binary_tree/lookup/BinaryTreeRandomLookup.kt
@@ -22,7 +22,7 @@ private class RandomGuide<PayloadT>(
 ) : Guide<PayloadT> {
     override fun instruct(
         payload: PayloadT,
-    ): Guide.Instruction = Guide.Instruction.Turn(
+    ): GuideInstruction = GuideInstruction.Turn(
         side = random.nextSide(),
     )
 }
