+++ DevToolkt/core/src/commonMain/kotlin/dev/toolkt/core/data_structures/binary_tree/lookup/BinaryTreeVolatileKeyLookup.kt
@@ -30,7 +30,7 @@ private class VolatileKeyOrderGuide<PayloadT, ColorT, KeyT : Comparable<KeyT>>(
 ) : Guide<PayloadT> {
     override fun instruct(
         payload: PayloadT,
-    ): Guide.Instruction {
+    ): GuideInstruction {
         val key = selector(payload)
 
         if (key == null) {
